Refactor pid.py to tools/pid.py. * Add fix for light sensor losing connection.
