+++ src/tools/exceptions.py
@@ -1,0 +1,3 @@
+class SensorException(Exception):
+    """Base class for sensor exceptions."""
+    pass
